// backup/src/types.tsx
export type TGenderScope = 'L' | 'P' | 'ALL';
export type TJurusanScope = 'KITAB' | 'TAHFIDZ' | 'ALL';

// Role sesuai request (huruf kecil)
export type TRole = 'super_admin' | 'rois' | 'bendahara' | 'kesantrian' | 'dewan' | 'alumni';
// ...
export interface IUserIdentity {
    id: string;
    name: string;
    avatar?: string;
    role: TRole;
    scopeGender: TGenderScope;
    scopeJurusan: TJurusanScope;
}